// src/enums/RoutesNames.ts
export enum RoutesNames {
  HOME = 'Home',
  BLOB_UTIL = 'Blob util',
  DEVICE_INFO = 'Device info',
  FAST_IMAGE = 'Fast image',
  HAPTIC_FEEDBACK = 'Haptic feedback',
  IMAGE_PICKER = 'Image picker',
  LINEAR_GRADIENT = 'Linear gradient',
  MMKV = 'MMKV',
  REANIMATED = 'ReAnimated',
  RN_SVG = 'React Native SVG',
  SKELETON_PLACEHOLDER = 'Skeleton placeholder',
  WEBVIEW = 'WEB view',
}